+++ CHANGELOG.md
@@ -5,3 +5,7 @@
 * removed signup possibility
 
 * check and fix sanofi on dev
+
+* if tool returns an array with first element having a file_id attribute, then the return is considered a list of chunks
+
+* collections, assistants, collections and files can now be made public. Only the owner can edit